+++ log/src/robocorp/log/_on_exit_context_manager.py
@@ -1,0 +1,9 @@
+class OnExitContextManager:
+    def __init__(self, on_exit):
+        self.on_exit = on_exit
+
+    def __enter__(self):
+        pass
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.on_exit()