+++ log/tests/robocorp_log_tests/test_impl_recovery.py
@@ -6,6 +6,8 @@ def test_impl_recovery_matches_suite() -> None:
     config.max_file_size_in_bytes = 9999999999
     config.max_files = 1
     config.log_html = None
+    config.min_messages_per_file = 10
+
     impl = _RoboOutputImpl(config)
     impl.start_run("suite1", 0)
     impl.start_task("test1", "modname", __file__, 0, "", 0)
@@ -25,6 +27,8 @@ def test_impl_recovery_matches_task() -> None:
     config.max_file_size_in_bytes = 9999999999
     config.max_files = 1
     config.log_html = None
+    config.min_messages_per_file = 10
+
     impl = _RoboOutputImpl(config)
     impl.start_run("suite1", 0)
     impl.start_task("test1", "modname", "source", 0, "", 0)
@@ -48,6 +52,8 @@ def test_impl_recovery_does_not_match_test() -> None:
     config.max_file_size_in_bytes = 9999999999
     config.max_files = 1
     config.log_html = None
+    config.min_messages_per_file = 10
+
     impl = _RoboOutputImpl(config)
     impl.start_run("suite1", 0)
     impl.start_task("test1", "modname", "source", 0, "", 0)
@@ -71,6 +77,8 @@ def test_impl_recovery_do_nothing() -> None:
     config.max_file_size_in_bytes = 9999999999
     config.max_files = 1
     config.log_html = None
+    config.min_messages_per_file = 10
+
     impl = _RoboOutputImpl(config)
     impl.end_run("suite1", "PASS", 0)
     assert len(impl._stack_handler._queue) == 0