+++ log/docs/changelog.md
@@ -1,11 +1,53 @@
-1.0.1
+NEXT (xxx-xx-xx)
+-----------------------------
+
+This release focused on improving the degenerate case where the chosen max log file
+was too close to the size of a single message.
+
+To improve this case, a `min_messages_per_file` was added to 
+`robocorp.log.add_log_output` where it'll only start rotating to a new file after
+a given amount of messages was given (the default is now `50`)
+and the maximum size for a repr can be set (by default it's now set to `200k` 
+chars before clipping the `repr(obj)`).
+
+- New API: `robocorp.log.setup_log`:
+
+```python
+def setup_log(*, max_value_repr_size: Optional[Union[str, int]] = None):
+    """
+    Setups the log "general" settings.
+
+    Args:
+        max_value_repr_size: This is the maximum number of chars which
+            may be used for a repr (values are clipped if a `repr(obj)` would
+            return a bigger representation).
+            May be passed directly as the value as an int or a string with the
+            value and associated unit.
+            Accepted units are: `k`, `m`.
+            Example: `"1000k"`, `"1m"`.
+
+            The default value for this setting is "200k".
+    """
+```
+
+- New parameter in `robocorp.log.add_log_output`:
+
+    `min_messages_per_file`: This is the minimum number of messages that need
+        to be added to a file for it to be rotated (if messages are too big
+        this may make the max_file_size be surpassed). This is needed to
+        prevent a case where a whole new file could be created after just
+        a single message if the message was too big for the max file size.
+
+
+1.0.1 (2023-06-14)
 -----------------------------
 
 - Improvements in `log.html`:
     - Variables are now pretty-printed in details.
     - Variables in the tree are always shown in a single line.
 
-1.0.0
+
+1.0.0 (2023-06-13)
 -----------------------------
 
 - `for` statement: code is no longer rewritten to add two `try..except..finally` blocks for each for loop.